Support compositing multiple transforms
The CSS properties scale, translate, rotate and transform each result in
the creation of a transform operation list.  Previously, the compositor
did not support having multiple keyframe models affecting the same
property.

In this patch, we defer the attachment of compositor keyframe models for
transforms until we have process all of the transform properties and
can arrange them in the proper order (translation, rotation, scale,
transform). When applying the keyframe models, we defer updating the
property trees until all of the transforms have been aggregated.

Bug: 696374
Change-Id: I279f5afc70ebc0cc72569d4d0d797699b753c4a6