Support BigInt/Uint64Array in crypto.getRandomValues
The check for "integer arrays" is to prevent people from using
getRandomValues with floating-point arrays, which doesn't work as people
might expect. However, the check was not updated when Chrome gained
support for BigInt64Array and BigUint64Array, which do work as expected.
This allows code fragments such as

    const b = new BigInt64Array(10);
    crypto.getRandomValues(b);

See https://github.com/w3c/webcrypto/issues/255 for the spec issue.

Bug: 1225765
Change-Id: I338ecc5594492e6f329580f4e8f04d367f866361